refactor!: simplify driving and driven port namespaces
