+++ src/Application/Bus/QueryDispatcher.php
@@ -14,7 +14,7 @@ namespace CloudCreativity\Modules\Application\Bus;
 
 use CloudCreativity\Modules\Contracts\Application\Bus\QueryHandlerContainer;
 use CloudCreativity\Modules\Contracts\Application\Messages\Query;
-use CloudCreativity\Modules\Contracts\Application\Ports\Driving\Queries\QueryDispatcher as IQueryDispatcher;
+use CloudCreativity\Modules\Contracts\Application\Ports\Driving\QueryDispatcher as IQueryDispatcher;
 use CloudCreativity\Modules\Contracts\Toolkit\Pipeline\PipeContainer;
 use CloudCreativity\Modules\Contracts\Toolkit\Result\Result;
 use CloudCreativity\Modules\Toolkit\Pipeline\MiddlewareProcessor;
