+++ src/Application/UnitOfWork/UnitOfWorkManager.php
@@ -13,8 +13,8 @@ declare(strict_types=1);
 namespace CloudCreativity\Modules\Application\UnitOfWork;
 
 use Closure;
-use CloudCreativity\Modules\Contracts\Application\Ports\Driven\Exceptions\ExceptionReporter;
-use CloudCreativity\Modules\Contracts\Application\Ports\Driven\UnitOfWork\UnitOfWork;
+use CloudCreativity\Modules\Contracts\Application\Ports\Driven\ExceptionReporter;
+use CloudCreativity\Modules\Contracts\Application\Ports\Driven\UnitOfWork;
 use CloudCreativity\Modules\Contracts\Application\UnitOfWork\UnitOfWorkManager as IUnitOfWorkManager;
 use RuntimeException;
 use Throwable;
